python/fastapi/myapp/requirements.txt
fastapi
fastapi_jwt_auth
pydantic==1.10.2
passlib
uvicorn
starlette
pydantic[email]
psycopg2-binary